Fix creating a new grid while a grid already exists
This update will fix drawing of new grids while a grid already exists.
Before this update, a new grid would be placed below the previous grid
and outside of the grid container.
By using a loop to remove all the children nodes from the grid
container, the new amount of children (squares) needed for the new grid
will be created and displayed inside of the grid container.